Add some Love to WidgetMatcher
Export MultiWidgetMatcher correctly
Move into separate file
Add API tests
Allow a widgetProp/elementProp to return null
