+++ lib/src/spot/diagnostic_props.dart
@@ -6,6 +6,7 @@ import 'package:meta/meta.dart';
 import 'package:spot/src/checks/checks_nullability.dart';
 import 'package:spot/src/spot/props.dart';
 import 'package:spot/src/spot/selectors.dart';
+import 'package:spot/src/spot/widget_matcher.dart';
 import 'package:spot/src/spot/widget_selector.dart';
 
 /// All [DiagnosticsProperty] related selectors
